requirements.txt
cycler==0.11.0
dill==0.3.4
EMD-signal==1.2.3
et-xmlfile==1.1.0
future==0.18.2
iso8601==1.0.2
kiwisolver==1.3.1
matplotlib==3.3.4
multiprocess==0.70.12.2
numpy==1.19.5
openpyxl==3.0.10
pandas==1.1.5
pathos==0.2.8
Pillow==8.4.0
pox==0.3.0
ppft==1.6.6.4
# pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.1.1
PyYAML==6.0
scipy==1.5.4
six==1.16.0
